Update functions.c
The single-cored bare medal program is basically done. I have run and tested it in my laptop but not my computer. The results are correct! The rest is to test it on the board and optimize the program. Note that when debuging the program I found the algorithm in the provided Haskell source code is flawed. It crop the a subimage with the size of 31*31. But this may not capture the whole pattern in the next image when the pattern moves 15 pixels in one frame. I fix the problem by increasing the size of  cropped subimage.